services/zoe-core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
redis==5.0.1
httpx==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
docker==6.1.3
psutil==5.9.6
anthropic==0.34.0
ollama==0.5.3
aiofiles==23.2.1
jinja2==3.1.2
requests==2.32.5
python-dotenv==1.0.0
openai>=1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
sentence-transformers>=5.1.0
livekit-api==0.8.2
faiss-cpu==1.7.4
numpy==1.24.3
# Light RAG dependencies
torch>=1.9.0
transformers>=4.21.0
scikit-learn>=1.0.0
huggingface-hub>=0.16.0
openai-whisper==20231117
schedule==1.2.0
litellm==1.43.7
prometheus-client==0.19.0
# Journal enhancement dependencies
Pillow>=10.0.0
pillow-heif>=0.13.0
geopy>=2.4.0
# memvid for learning archives (Phase 6)
memvid>=0.1.3
qrcode>=8.0
PyPDF2>=3.0.0
opencv-python>=4.8.0
# Push notifications (Phase 2)
py-vapid>=1.9.1
pywebpush>=2.0.0
